Extract QueryRunnerToSqlPlanNodeVisitor to separate files
