+++ velox/exec/fuzzer/PrestoQueryRunner.cpp
@@ -28,6 +28,7 @@
 #include "velox/dwio/dwrf/writer/Writer.h"
 #include "velox/exec/fuzzer/FuzzerUtil.h"
 #include "velox/exec/fuzzer/PrestoQueryRunner.h"
+#include "velox/exec/fuzzer/PrestoQueryRunnerToSqlPlanNodeVisitor.h"
 #include "velox/exec/fuzzer/PrestoSql.h"
 #include "velox/exec/tests/utils/QueryAssertions.h"
 #include "velox/functions/prestosql/types/IPAddressType.h"
@@ -41,507 +42,6 @@ using namespace facebook::velox;
 
 namespace facebook::velox::exec::test {
 namespace {
-std::string toWindowCallSql(
-    const core::CallTypedExprPtr& call,
-    bool ignoreNulls = false) {
-  std::stringstream sql;
-  sql << call->name() << "(";
-  toCallInputsSql(call->inputs(), sql);
-  sql << ")";
-  if (ignoreNulls) {
-    sql << " IGNORE NULLS";
-  }
-  return sql.str();
-}
-
-class PrestoQueryRunnerToSqlPlanNodeVisitor : public PrestoSqlPlanNodeVisitor {
- public:
-  PrestoQueryRunnerToSqlPlanNodeVisitor(
-      PrestoQueryRunner* queryRunner,
-      const std::shared_ptr<QueryRunnerContext>& queryRunnerContext_)
-      : PrestoSqlPlanNodeVisitor(queryRunner),
-        queryRunnerContext_(queryRunnerContext_) {}
-
-  void visit(
-      const core::AggregationNode& node,
-      core::PlanNodeVisitorContext& ctx) const override {
-    // Assume plan is Aggregation over Values.
-    VELOX_CHECK(node.step() == core::AggregationNode::Step::kSingle);
-
-    PrestoSqlPlanNodeVisitorContext& visitorContext =
-        static_cast<PrestoSqlPlanNodeVisitorContext&>(ctx);
-
-    if (!queryRunner_->isSupportedDwrfType(node.sources()[0]->outputType())) {
-      visitorContext.sql = std::nullopt;
-      return;
-    }
-
-    std::vector<std::string> groupingKeys;
-    for (const auto& key : node.groupingKeys()) {
-      groupingKeys.push_back(key->name());
-    }
-
-    std::stringstream sql;
-    sql << "SELECT " << folly::join(", ", groupingKeys);
-
-    const auto& aggregates = node.aggregates();
-    if (!aggregates.empty()) {
-      if (!groupingKeys.empty()) {
-        sql << ", ";
-      }
-
-      for (auto i = 0; i < aggregates.size(); ++i) {
-        appendComma(i, sql);
-        const auto& aggregate = aggregates[i];
-        sql << toAggregateCallSql(
-            aggregate.call,
-            aggregate.sortingKeys,
-            aggregate.sortingOrders,
-            aggregate.distinct);
-
-        if (aggregate.mask != nullptr) {
-          sql << " filter (where " << aggregate.mask->name() << ")";
-        }
-        sql << " as " << node.aggregateNames()[i];
-      }
-    }
-    // AggregationNode should have a single source.
-    const auto source = toSql(node.sources()[0]);
-    if (!source) {
-      visitorContext.sql = std::nullopt;
-      return;
-    }
-    sql << " FROM " << *source;
-
-    if (!groupingKeys.empty()) {
-      sql << " GROUP BY " << folly::join(", ", groupingKeys);
-    }
-
-    visitorContext.sql = sql.str();
-  }
-
-  void visit(const core::ArrowStreamNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::AssignUniqueIdNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::EnforceSingleRowNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::ExchangeNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::ExpandNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::FilterNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::GroupIdNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::HashJoinNode& node, core::PlanNodeVisitorContext& ctx)
-      const override {
-    PrestoSqlPlanNodeVisitor::visit(node, ctx);
-  }
-
-  void visit(const core::IndexLookupJoinNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::LimitNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::LocalMergeNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::LocalPartitionNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::MarkDistinctNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::MergeExchangeNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::MergeJoinNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(
-      const core::NestedLoopJoinNode& node,
-      core::PlanNodeVisitorContext& ctx) const override {
-    PrestoSqlPlanNodeVisitor::visit(node, ctx);
-  }
-
-  void visit(const core::OrderByNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::PartitionedOutputNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::ProjectNode& node, core::PlanNodeVisitorContext& ctx)
-      const override {
-    PrestoSqlPlanNodeVisitorContext& visitorContext =
-        static_cast<PrestoSqlPlanNodeVisitorContext&>(ctx);
-
-    const auto sourceSql = toSql(node.sources()[0]);
-    if (!sourceSql.has_value()) {
-      visitorContext.sql = std::nullopt;
-      return;
-    }
-
-    std::stringstream sql;
-    sql << "SELECT ";
-
-    for (auto i = 0; i < node.names().size(); ++i) {
-      appendComma(i, sql);
-      auto projection = node.projections()[i];
-      if (auto field =
-              std::dynamic_pointer_cast<const core::FieldAccessTypedExpr>(
-                  projection)) {
-        if (field->isInputColumn()) {
-          sql << field->name();
-        } else {
-          toCallInputsSql(field->inputs(), sql);
-          sql << fmt::format(".{}", field->name());
-        }
-      } else if (
-          auto call = std::dynamic_pointer_cast<const core::CallTypedExpr>(
-              projection)) {
-        sql << toCallSql(call);
-      } else if (
-          auto cast = std::dynamic_pointer_cast<const core::CastTypedExpr>(
-              projection)) {
-        sql << toCastSql(*cast);
-      } else if (
-          auto concat = std::dynamic_pointer_cast<const core::ConcatTypedExpr>(
-              projection)) {
-        sql << toConcatSql(*concat);
-      } else if (
-          auto constant =
-              std::dynamic_pointer_cast<const core::ConstantTypedExpr>(
-                  projection)) {
-        sql << toConstantSql(*constant);
-      } else {
-        VELOX_NYI();
-      }
-
-      sql << " as " << node.names()[i];
-    }
-
-    sql << " FROM (" << sourceSql.value() << ")";
-    visitorContext.sql = sql.str();
-  }
-
-  void visit(const core::RowNumberNode& node, core::PlanNodeVisitorContext& ctx)
-      const override {
-    PrestoSqlPlanNodeVisitorContext& visitorContext =
-        static_cast<PrestoSqlPlanNodeVisitorContext&>(ctx);
-
-    if (!queryRunner_->isSupportedDwrfType(node.sources()[0]->outputType())) {
-      visitorContext.sql = std::nullopt;
-      return;
-    }
-
-    std::stringstream sql;
-    sql << "SELECT ";
-
-    const auto& inputType = node.sources()[0]->outputType();
-    for (auto i = 0; i < inputType->size(); ++i) {
-      appendComma(i, sql);
-      sql << inputType->nameOf(i);
-    }
-
-    sql << ", row_number() OVER (";
-
-    const auto& partitionKeys = node.partitionKeys();
-    if (!partitionKeys.empty()) {
-      sql << "partition by ";
-      for (auto i = 0; i < partitionKeys.size(); ++i) {
-        appendComma(i, sql);
-        sql << partitionKeys[i]->name();
-      }
-    }
-
-    // RowNumberNode should have a single source.
-    const auto source = toSql(node.sources()[0]);
-    if (!source) {
-      visitorContext.sql = std::nullopt;
-      return;
-    }
-    sql << ") as row_number FROM " << *source;
-
-    visitorContext.sql = sql.str();
-  }
-
-  void visit(const core::TableScanNode& node, core::PlanNodeVisitorContext& ctx)
-      const override {
-    PrestoSqlPlanNodeVisitor::visit(node, ctx);
-  }
-
-  void visit(
-      const core::TableWriteNode& node,
-      core::PlanNodeVisitorContext& ctx) const override {
-    PrestoSqlPlanNodeVisitorContext& visitorContext =
-        static_cast<PrestoSqlPlanNodeVisitorContext&>(ctx);
-
-    auto insertTableHandle =
-        std::dynamic_pointer_cast<connector::hive::HiveInsertTableHandle>(
-            node.insertTableHandle()->connectorInsertTableHandle());
-
-    // Returns a CTAS sql with specified table properties from TableWriteNode,
-    // example sql:
-    // CREATE TABLE tmp_write WITH (
-    // PARTITIONED_BY = ARRAY['p0'],
-    // BUCKETED_COUNT = 2, BUCKETED_BY = ARRAY['b0', 'b1'],
-    // SORTED_BY = ARRAY['s0 ASC', 's1 DESC'],
-    // FORMAT = 'ORC'
-    // )
-    // AS SELECT * FROM t_<id>
-    std::stringstream sql;
-    sql << "CREATE TABLE tmp_write";
-    std::vector<std::string> partitionKeys;
-    for (auto i = 0; i < node.columnNames().size(); ++i) {
-      if (insertTableHandle->inputColumns()[i]->isPartitionKey()) {
-        partitionKeys.push_back(insertTableHandle->inputColumns()[i]->name());
-      }
-    }
-    sql << " WITH (";
-
-    if (insertTableHandle->isPartitioned()) {
-      sql << " PARTITIONED_BY = ARRAY[";
-      for (int i = 0; i < partitionKeys.size(); ++i) {
-        appendComma(i, sql);
-        sql << "'" << partitionKeys[i] << "'";
-      }
-      sql << "], ";
-
-      if (insertTableHandle->bucketProperty() != nullptr) {
-        const auto bucketCount =
-            insertTableHandle->bucketProperty()->bucketCount();
-        const auto& bucketColumns =
-            insertTableHandle->bucketProperty()->bucketedBy();
-        sql << " BUCKET_COUNT = " << bucketCount << ", BUCKETED_BY = ARRAY[";
-        for (int i = 0; i < bucketColumns.size(); ++i) {
-          appendComma(i, sql);
-          sql << "'" << bucketColumns[i] << "'";
-        }
-        sql << "], ";
-
-        const auto& sortColumns =
-            insertTableHandle->bucketProperty()->sortedBy();
-        if (!sortColumns.empty()) {
-          sql << " SORTED_BY = ARRAY[";
-          for (int i = 0; i < sortColumns.size(); ++i) {
-            appendComma(i, sql);
-            sql << "'" << sortColumns[i]->sortColumn() << " "
-                << (sortColumns[i]->sortOrder().isAscending() ? "ASC" : "DESC")
-                << "'";
-          }
-          sql << "], ";
-        }
-      }
-    }
-
-    // TableWriteNode should have a single source.
-    const auto source = toSql(node.sources()[0]);
-    if (!source) {
-      visitorContext.sql = std::nullopt;
-      return;
-    }
-    sql << "FORMAT = 'ORC')  AS SELECT * FROM " << *source;
-
-    visitorContext.sql = sql.str();
-  }
-
-  void visit(const core::TableWriteMergeNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::TopNNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(
-      const core::TopNRowNumberNode& node,
-      core::PlanNodeVisitorContext& ctx) const override {
-    PrestoSqlPlanNodeVisitorContext& visitorContext =
-        static_cast<PrestoSqlPlanNodeVisitorContext&>(ctx);
-
-    if (!queryRunner_->isSupportedDwrfType(node.sources()[0]->outputType())) {
-      visitorContext.sql = std::nullopt;
-      return;
-    }
-
-    std::stringstream sql;
-    sql << "SELECT * FROM (SELECT ";
-
-    const auto& inputType = node.sources()[0]->outputType();
-    for (auto i = 0; i < inputType->size(); ++i) {
-      appendComma(i, sql);
-      sql << inputType->nameOf(i);
-    }
-
-    sql << ", row_number() OVER (";
-
-    const auto& partitionKeys = node.partitionKeys();
-    if (!partitionKeys.empty()) {
-      sql << "partition by ";
-      for (auto i = 0; i < partitionKeys.size(); ++i) {
-        appendComma(i, sql);
-        sql << partitionKeys[i]->name();
-      }
-    }
-
-    const auto& sortingKeys = node.sortingKeys();
-    const auto& sortingOrders = node.sortingOrders();
-
-    if (!sortingKeys.empty()) {
-      sql << " ORDER BY ";
-      for (auto j = 0; j < sortingKeys.size(); ++j) {
-        appendComma(j, sql);
-        sql << sortingKeys[j]->name() << " " << sortingOrders[j].toString();
-      }
-    }
-
-    std::string rowNumberColumnName = node.generateRowNumber()
-        ? node.outputType()->nameOf(node.outputType()->children().size() - 1)
-        : "row_number";
-
-    // TopNRowNumberNode should have a single source.
-    const auto source = toSql(node.sources()[0]);
-    if (!source) {
-      visitorContext.sql = std::nullopt;
-      return;
-    }
-    sql << ") as " << rowNumberColumnName << " FROM " << *source << ") ";
-    sql << " where " << rowNumberColumnName << " <= " << node.limit();
-
-    visitorContext.sql = sql.str();
-  }
-
-  void visit(const core::TraceScanNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::UnnestNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
-  void visit(const core::ValuesNode& node, core::PlanNodeVisitorContext& ctx)
-      const override {
-    PrestoSqlPlanNodeVisitor::visit(node, ctx);
-  }
-
-  void visit(const core::WindowNode& node, core::PlanNodeVisitorContext& ctx)
-      const override {
-    PrestoSqlPlanNodeVisitorContext& visitorContext =
-        static_cast<PrestoSqlPlanNodeVisitorContext&>(ctx);
-
-    if (!queryRunner_->isSupportedDwrfType(node.sources()[0]->outputType())) {
-      visitorContext.sql = std::nullopt;
-      return;
-    }
-
-    std::stringstream sql;
-    sql << "SELECT ";
-
-    const auto& inputType = node.sources()[0]->outputType();
-    for (auto i = 0; i < inputType->size(); ++i) {
-      appendComma(i, sql);
-      sql << inputType->nameOf(i);
-    }
-
-    sql << ", ";
-
-    const auto& functions = node.windowFunctions();
-    for (auto i = 0; i < functions.size(); ++i) {
-      appendComma(i, sql);
-      sql << toWindowCallSql(
-          functions[i].functionCall, functions[i].ignoreNulls);
-
-      sql << " OVER (";
-
-      const auto& partitionKeys = node.partitionKeys();
-      if (!partitionKeys.empty()) {
-        sql << "PARTITION BY ";
-        for (auto j = 0; j < partitionKeys.size(); ++j) {
-          appendComma(j, sql);
-          sql << partitionKeys[j]->name();
-        }
-      }
-
-      const auto& sortingKeys = node.sortingKeys();
-      const auto& sortingOrders = node.sortingOrders();
-
-      if (!sortingKeys.empty()) {
-        sql << " ORDER BY ";
-        for (auto j = 0; j < sortingKeys.size(); ++j) {
-          appendComma(j, sql);
-          sql << sortingKeys[j]->name() << " " << sortingOrders[j].toString();
-        }
-      }
-
-      sql << " " << queryRunnerContext_->windowFrames_.at(node.id()).at(i);
-      sql << ")";
-    }
-
-    // WindowNode should have a single source.
-    const auto source = toSql(node.sources()[0]);
-    if (!source) {
-      visitorContext.sql = std::nullopt;
-      return;
-    }
-    sql << " FROM " << *source;
-
-    visitorContext.sql = sql.str();
-  }
-
-  /// Used to visit custom PlanNodes that extend the set provided by Velox.
-  void visit(const core::PlanNode&, core::PlanNodeVisitorContext&)
-      const override {
-    VELOX_NYI();
-  }
-
- private:
-  std::shared_ptr<QueryRunnerContext> queryRunnerContext_;
-};
 
 void writeToFile(
     const std::string& path,
@@ -664,7 +164,7 @@ PrestoQueryRunner::PrestoQueryRunner(
 std::optional<std::string> PrestoQueryRunner::toSql(
     const core::PlanNodePtr& plan) {
   PrestoSqlPlanNodeVisitorContext context;
-  PrestoQueryRunnerToSqlPlanNodeVisitor visitor(this, queryRunnerContext_);
+  PrestoQueryRunnerToSqlPlanNodeVisitor visitor(queryRunnerContext_);
   plan->accept(visitor, context);
 
   return context.sql;
@@ -684,6 +184,21 @@ const std::vector<TypePtr>& PrestoQueryRunner::supportedScalarTypes() const {
       TIMESTAMP(),
   };
   return kScalarTypes;
+}
+
+// static
+bool PrestoQueryRunner::isSupportedDwrfType(const TypePtr& type) {
+  if (type->isDate() || type->isIntervalDayTime() || type->isUnKnown()) {
+    return false;
+  }
+
+  for (auto i = 0; i < type->size(); ++i) {
+    if (!isSupportedDwrfType(type->childAt(i))) {
+      return false;
+    }
+  }
+
+  return true;
 }
 
 const std::unordered_map<std::string, DataSpec>&
