+++ velox/exec/fuzzer/PrestoSql.h
@@ -19,8 +19,6 @@
 
 namespace facebook::velox::exec::test {
 
-class ReferenceQueryRunner;
-
 /// Appends a comma to a given stringstream if the provided integer is greater
 /// than 0.
 void appendComma(int32_t i, std::stringstream& sql);
@@ -63,8 +61,7 @@ class PrestoSqlPlanNodeVisitorContext : public core::PlanNodeVisitorContext {
 
 class PrestoSqlPlanNodeVisitor : public core::PlanNodeVisitor {
  public:
-  explicit PrestoSqlPlanNodeVisitor(ReferenceQueryRunner* queryRunner)
-      : queryRunner_(queryRunner) {}
+  explicit PrestoSqlPlanNodeVisitor() {}
 
   void visit(const core::HashJoinNode& node, core::PlanNodeVisitorContext& ctx)
       const override;
@@ -81,7 +78,5 @@ class PrestoSqlPlanNodeVisitor : public core::PlanNodeVisitor {
 
  protected:
   std::optional<std::string> toSql(const core::PlanNodePtr& node) const;
-
-  ReferenceQueryRunner* queryRunner_;
 };
 } // namespace facebook::velox::exec::test
